requirements.txt
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
python-dotenv==0.15.0
